- Implement GPU buffer source for processing (OpenCL, OpenGL, DirectX currently) - Implement drawing in the GPU kernels - Fix zero-copy Qt RHI pipeline - Add "Processing resolution" in Synchronization - Add support for lens interpolation (Hero 11 Autoboost mode and zoom lenses on supported cameras (eg. Sony)) - Rewrite "digital lens" support, make it possible to handle different processed lens types (eg. GoPro Superview, Hyperview, Stretch) - Log panics to gyroflow.log file - Reduce mutex locking - Move Video speed to "Stabilization" section

+++ src/core/zooming/fov_iterative.rs
@@ -18,9 +18,9 @@ Iterative FOV calculation:
 */
 
 pub struct FovIterative {
-    input_dim: (f64, f64),
-    output_dim: (f64, f64),
-    output_inv_aspect: f64,
+    input_dim: (f32, f32),
+    output_dim: (f32, f32),
+    output_inv_aspect: f32,
     compute_params: ComputeParams,
     debug_points: RwLock<BTreeMap<i64, Vec<(f64, f64)>>>,
 }
@@ -65,9 +65,9 @@ impl FieldOfViewAlgorithm for FovIterative {
 
 impl FovIterative {
     pub fn new(compute_params: ComputeParams) -> Self {
-        let ratio = compute_params.video_width as f64 / compute_params.video_output_width.max(1) as f64;
-        let input_dim = (compute_params.video_width as f64, compute_params.video_height as f64);
-        let output_dim = (compute_params.video_output_width as f64 * ratio, compute_params.video_output_height as f64 * ratio);
+        let ratio = compute_params.video_width as f32 / compute_params.video_output_width.max(1) as f32;
+        let input_dim = (compute_params.video_width as f32, compute_params.video_height as f32);
+        let output_dim = (compute_params.video_output_width as f32 * ratio, compute_params.video_output_height as f32 * ratio);
         let output_inv_aspect = output_dim.1 / output_dim.0;
 
         Self {
@@ -79,7 +79,7 @@ impl FovIterative {
         }
     }
 
-    fn find_fov(&self, rect: &[(f64, f64)], ts: f64, center: &Point2D) -> f64 {
+    fn find_fov(&self, rect: &[(f32, f32)], ts: f64, center: &Point2D) -> f64 {
         let ts_us = (ts * 1000.0).round() as i64;
 
         let adaptive_zoom_center_x = self.compute_params.keyframes.value_at_video_timestamp(&KeyframeType::ZoomingCenterX, ts).unwrap_or(self.compute_params.adaptive_zoom_center_offset.0);
@@ -87,11 +87,11 @@ impl FovIterative {
 
         let mut polygon = undistort_points_with_rolling_shutter(&rect, ts, &self.compute_params);
         for (x, y) in polygon.iter_mut() {
-            *x -= adaptive_zoom_center_x * self.input_dim.0;
-            *y -= adaptive_zoom_center_y * self.input_dim.1;
+            *x -= adaptive_zoom_center_x as f32 * self.input_dim.0;
+            *y -= adaptive_zoom_center_y as f32 * self.input_dim.1;
         }
         if self.compute_params.zooming_debug_points {
-            self.debug_points.write().insert(ts_us, polygon.iter().map(|(x, y)| (x / self.input_dim.0, y / self.input_dim.1)).collect());
+            self.debug_points.write().insert(ts_us, polygon.iter().map(|(x, y)| ((x / self.input_dim.0) as f64, (y / self.input_dim.1) as f64)).collect());
         }
 
         let initial = (1000000.0, 1000000.0 * self.output_inv_aspect);
@@ -111,8 +111,8 @@ impl FovIterative {
                 let distorted = interpolate_points(&relevant, 30);
                 polygon = undistort_points_with_rolling_shutter(&distorted, ts, &self.compute_params);
                 for (x, y) in polygon.iter_mut() {
-                    *x -= adaptive_zoom_center_x * self.input_dim.0;
-                    *y -= adaptive_zoom_center_y * self.input_dim.1;
+                    *x -= adaptive_zoom_center_x as f32 * self.input_dim.0;
+                    *y -= adaptive_zoom_center_y as f32 * self.input_dim.1;
                 }
                 nearest = self.nearest_edge(&polygon, center, nearest.1);
             } else {
@@ -120,10 +120,10 @@ impl FovIterative {
             }
         }
 
-        nearest.1.0 * 2.0 / self.output_dim.0
+        (nearest.1.0 * 2.0 / self.output_dim.0) as f64
     }
 
-    fn nearest_edge(&self, polygon: &[(f64, f64)], center: &Point2D, initial: (f64, f64)) -> (Option<usize>,(f64, f64)) {
+    fn nearest_edge(&self, polygon: &[(f32, f32)], center: &Point2D, initial: (f32, f32)) -> (Option<usize>, (f32, f32)) {
         polygon
             .iter()
             .enumerate()
@@ -142,20 +142,20 @@ impl FovIterative {
 }
 
 // Returns points placed around a rectangle in a continous order
-fn points_around_rect(mut w: f64, mut h: f64, w_div: usize, h_div: usize) -> Vec<(f64, f64)> {
+fn points_around_rect(mut w: f32, mut h: f32, w_div: usize, h_div: usize) -> Vec<(f32, f32)> {
     let margin = 2.0;
     w -= margin * 2.0;
     h -= margin * 2.0;
 
     let (wcnt, hcnt) = (w_div.max(2) - 1, h_div.max(2) - 1);
-    let (wstep, hstep) = (w / wcnt as f64, h / hcnt as f64);
+    let (wstep, hstep) = (w / wcnt as f32, h / hcnt as f32);
 
     // ordered!
-    let mut distorted_points: Vec<(f64, f64)> = Vec::with_capacity((wcnt + hcnt) * 2);
-    for i in 0..wcnt { distorted_points.push((i as f64 * wstep,          0.0)); }
-    for i in 0..hcnt { distorted_points.push((w,                         i as f64 * hstep)); }
-    for i in 0..wcnt { distorted_points.push(((wcnt - i) as f64 * wstep, h)); }
-    for i in 0..hcnt { distorted_points.push((0.0,                       (hcnt - i) as f64 * hstep)); }
+    let mut distorted_points: Vec<(f32, f32)> = Vec::with_capacity((wcnt + hcnt) * 2);
+    for i in 0..wcnt { distorted_points.push((i as f32 * wstep,          0.0)); }
+    for i in 0..hcnt { distorted_points.push((w,                         i as f32 * hstep)); }
+    for i in 0..wcnt { distorted_points.push(((wcnt - i) as f32 * wstep, h)); }
+    for i in 0..hcnt { distorted_points.push((0.0,                       (hcnt - i) as f32 * hstep)); }
 
     // Add margin
     for (x, y) in distorted_points.iter_mut() {
@@ -167,13 +167,13 @@ fn points_around_rect(mut w: f64, mut h: f64, w_div: usize, h_div: usize) -> Vec
 }
 
 // linear interpolates steps between points in array
-fn interpolate_points(pts: &[(f64, f64)], steps: usize) -> Vec<(f64,f64)> {
+fn interpolate_points(pts: &[(f32, f32)], steps: usize) -> Vec<(f32,f32)> {
     let d = steps+1;
     let new_len = d * pts.len() - steps;
     (0..new_len).map(|i| {
         let idx1 = i / d;
         let idx2 = (idx1+1).min(pts.len()-1);
-        let f = ((i % d) as f64) / (d as f64);
+        let f = ((i % d) as f32) / (d as f32);
         (pts[idx1].0 + f * (pts[idx2].0 - pts[idx1].0), pts[idx1].1 + f * (pts[idx2].1 - pts[idx1].1))
     }).collect()
 }
